Swap backend from `emacs-web-server` to `elnode`
Emacs web server provides no method for extracting the content of a post
request. Since "application/json" is not a supported content type, there appears
to be no way to post JSON data. You can't access the content to decode it
yourself.

Initially, I was using a hack to manually extract the body from the request, but
this hack was unreliable. It only worked around 70% of the time.

Until this is fixed, use Elnode. Elnode has its own issues (a manual patch needs
to be shipped with the package to make Elnode functional on Emacs 26+), but I
can reliably extract the body of POST requests with it.

The old `emacs-web-server` method is still largely complete - this is the only
large issue. Thus it should be separated onto its own branch and stored for easy
reimplementation at a later date. Ensure the POST body issue is clearly
explained - best to throw an error when extracting it so there is no ambiguity.