Change architecture to allow multiple independent RPC servers
This is a large commit. All tests pass afterwards.

Change the architecture of the package so each server operates independently.
Each server will have its own list of exposed functions and its own username and
password. Each server will be assigned a different dynamic port and this
information will be stored in a user-only temporary directory specific to that
server.

The entry point to the package is modified so most servers will be automatically
assigned a port, username and password at creation. This information is written
to a session file with a predictable path so clients can set up connections
automatically. Permissions for this file should be restricted to the current
user.

Advanced server creation allows the user finer control over this behavior. For
example, they can specify the port, credentials and what will be published.

This allows many different RPC packages to be written, each running on a
different server, without interfering with one another (although
process-intensive RPC calls will still block the main thread, thus the ability
to serve other clients). The purpose of this package is to act as a framework
that allows client/server tools to be built on top of Emacs. Each package needs
to be confident that it's segregated.

The README is *not* updated with this change. The intention in the long-run is
to rename the package. The README will be overhauled once that's been done.