Replace handler ending system - use catch/throw instead of errors
Porthole handlers were terminated with a special class of errors. These errors
were caught at the top of the handler, and used to construct a response. The
Emacs manual explicitly discourages using errors like this (to prematurely
terminate functions). It recommends `catch` and `throw`.

Switch to a catching system, where the details of the response to be sent are
thrown to the top of the handler. The handler can catch these, send a response,
and immediately terminate. Wrap this throwing system in a number of convenience
methods, to make it intuitive for the developer.