Replace Elnode with emacs-web-server
Giant commit - can't easily swap out part of the backend so this is a complete
replacement.

Elnode was causing a lot of problems. Logging by default, unstable connections,
Emacs would hang for a long time when individual connections failed. Temp
buffers change the state during RPC calls, and seem slow.

Instead, I've swapped the backend over to `web-server`. It appears faster, and
much more robust. It also cuts down necessary code. `web-server` also appears to
have websocket capabilities, which may come in useful in the future. Persistent
connections would be nice.

This replacement relies on being able to access POST content of requests
directly. This is not currently possible in the official `web-server` - it
relies on a PR I've made on GitHub (currently pending). This backend will have
to wait for that PR, unless I directly vendor the modified `web-server` as a
porthole submodule.

All tests pass with the new implementation. Some tests have been modified.